Prepare API for batching; remove deprecated .NET 7 and Xamarin
