+++ .github/workflows/githubpages.yml
@@ -17,12 +17,12 @@ jobs:
         uses: actions/checkout@main
         with:
           fetch-depth: 0
-      - name: Setup .NET 8
+      - name: Setup .NET
         uses: actions/setup-dotnet@main
         with:
           dotnet-version: 8.x
-      - name: Setup .NET workload android
-        run: dotnet workload install android
+      - name: Setup .NET workload maui-android
+        run: dotnet workload install maui-android
       - name: Setup .NET workload wasm-tools
         run: dotnet workload install wasm-tools
       - name: Update relative paths
