+++ .github/workflows/githubpages_staging.yml
@@ -18,12 +18,12 @@ jobs:
         uses: actions/checkout@main
         with:
           fetch-depth: 0
-      - name: Setup .NET 8
+      - name: Setup .NET
         uses: actions/setup-dotnet@main
         with:
           dotnet-version: 8.x
-      - name: Setup .NET workload android
-        run: dotnet workload install android
+      - name: Setup .NET workload maui-android
+        run: dotnet workload install maui-android
       - name: Setup .NET workload wasm-tools
         run: dotnet workload install wasm-tools
       - name: Update relative paths
@@ -79,6 +79,10 @@ jobs:
     needs: publish
     runs-on: ubuntu-latest
 
+    concurrency:
+      group: "pages-stating"
+      cancel-in-progress: true
+
     # Grant GITHUB_TOKEN the permissions required to make a Pages deployment
     permissions:
       pages: write      # to deploy to Pages
@@ -87,10 +91,6 @@ jobs:
     environment:
       name: github-pages-staging
       url: ${{ steps.deployment.outputs.page_url }}
-
-    concurrency:
-      group: "pages"
-      cancel-in-progress: true
     
     steps:
       - name: Deploy to GitHub Pages
