+++ src/PDFtoImage/Internals/IsExternalInit.cs
@@ -1,4 +1,4 @@
-﻿#if NETSTANDARD || MONOANDROID || NETFRAMEWORK
+﻿#if NETSTANDARD || NETFRAMEWORK
 namespace System.Runtime.CompilerServices
 {
     internal static class IsExternalInit { }
